refactor new roles table
